프로젝트 로컬 원본/프로젝트 로컬 원본/projectback/requirements.txt
aioredis==1.3.1
api==0.0.7
APScheduler==3.6.3
asgiref==2.3.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
beautifulsoup4==4.8.2
bootstrap4==0.1.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.3.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
daphne==2.4.1
disqus==0.0.4
Django==2.2.10
django-appconf==1.0.3
django-bootstrap-pagination==1.7.1
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-imagekit==4.0.2
django-ip==1.0.2
django-pagination==1.0.7
django-pagination-bootstrap==1.3
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
ffmpeg==1.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
image==1.5.27
imageio==2.6.1
incremental==17.5.0
ipaddr==2.2.0
isort==4.3.21
kiwisolver==1.1.0
kma==0.3.0
lazy-object-proxy==1.4.3
lxml==4.5.0
matplotlib==3.0.3
mccabe==0.6.1
msgpack==0.5.6
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nose==1.3.7
np==1.0.2
npm==0.1.1
numpy==1.18.1
optional-django==0.1.0
pandas==0.25.3
pilkit==2.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.7
wrapt==1.11.2
zope.interface==4.7.1
